+++ gen/templates/profile_parcels/data-source.tf
@@ -1,0 +1,8 @@
+data "sdwan_{{snakeCase .Name}}_profile_parcel" "example" {
+  id = "f6b2c44c-693c-4763-b010-895aa3d236bd"
+{{- range  .Attributes}}
+{{- if .Reference}}
+  {{.TfName}} = {{if eq .Type "String"}}"{{.Example}}"{{else if eq .Type "StringList"}}["{{.Example}}"]{{else}}{{.Example}}{{end}}
+{{- end}}
+{{- end}}
+}